Documentation and interface format changes
I added additional documentation to the files and individual methods.
In addition, I added an entry to the .gitignore file, in case this tool
is expanded in the future to utilize additional python files inside the
same directory.

Finally, I reworked the interface of this tool so that it will be run from
the command line with a flag and a url that indicate what types of information
it should gatehr and from what site.